+++ view/view_office.php
@@ -9,7 +9,7 @@
     <script src="https://cdn.jsdelivr.net/npm/@tailwindcss/browser@4"></script>
 </head>
 
-<body>
+<body class="bg-zinc-300">
     <div class="container mx-auto p-6">
         <h1 class="text-3xl font-bold text-gray-800 mb-6 text-center">Office</h1>
         <div class="overflow-x-auto shadow-lg rounded-lg">
@@ -37,7 +37,7 @@
                         <td class="px-6 py-4 whitespace-nowrap text-sm text-gray-900"><?=$office->getTelepon()?></td>
                         <td class="px-6 py-4 whitespace-nowrap text-sm font-medium">
                             <!-- <a href="#" class="text-indigo-600 hover:text-indigo-900 mr-3">Edit</a> -->
-                            <a href="../controller/controller_member.php?deleteID=<?=$index?>" class="text-red-600 hover:text-red-900">Delete</a>
+                            <a href="../controller/controller_member.php?deleteID=<?=$index?>" class="text-red-600 hover:text-red-900" onclick="return confirm('Do you want to delete this office?');">Delete</a>
                         </td>
                     </tr>
                     <?php
@@ -48,13 +48,13 @@
         </div>
         
         <div class="flex justify-center mt-6">
-            <a href="view_addoffice.php" class="bg-green-500 hover:bg-green-600 text-white font-semibold py-3 px-6 rounded-lg transition duration-200">
+            <a href="view_addoffice.php" class="bg-zinc-500 hover:bg-blue-600 text-white font-semibold py-3 px-6 rounded-lg transition duration-200">
                 Add Office
             </a>
         </div>
 
         <div class="flex justify-center mt-4">
-            <a href="../index.php" class="bg-blue-500 hover:bg-blue-600 text-white font-semibold py-3 px-6 rounded-lg transition duration-200">
+            <a href="../index.php" class="bg-zinc-500 hover:bg-blue-600 text-white font-semibold py-3 px-6 rounded-lg transition duration-200">
                 Back to Home
             </a>
         </div>
